backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
djoser==2.3.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-dotenv==1.1.0
django-filter==25.1
drf-extra-fields==3.7.0
gunicorn==23.0.0
Pillow==11.2.1